Added the OOSystemDescriptionManager files to the mac port of the project

+++ src/Core/OOSystemDescriptionManager.m
@@ -306,7 +306,7 @@ static NSString *kOOSystemLayerProperty = @"layer";
 		NSUInteger index = ([PLAYER galaxyNumber] * OO_SYSTEMS_PER_GALAXY) + s;
 		if (index >= OO_SYSTEM_CACHE_LENGTH)
 		{
-			OOLog(@"system.description.error",@"'%u' is an invalid system index for the current system. This is an internal error. Please report it.",index);
+			OOLog(@"system.description.error",@"'%lu' is an invalid system index for the current system. This is an internal error. Please report it.",index);
 			return [NSDictionary dictionary];
 		}
 		return propertyCache[index];
@@ -503,7 +503,7 @@ static NSString *kOOSystemLayerProperty = @"layer";
 - (void) updateCacheEntry:(NSUInteger)i
 {
 	NSAssert(i < OO_SYSTEM_CACHE_LENGTH,@"Invalid cache entry number");
-	NSString *key = [NSString stringWithFormat:@"%u %u",i/OO_SYSTEMS_PER_GALAXY,i%OO_SYSTEMS_PER_GALAXY];
+	NSString *key = [NSString stringWithFormat:@"%lu %lu",i/OO_SYSTEMS_PER_GALAXY,i%OO_SYSTEMS_PER_GALAXY];
 	NSDictionary *current = [self calculatePropertiesForSystemKey:key];
 
 	[propertyCache[i] removeAllObjects];
@@ -514,7 +514,7 @@ static NSString *kOOSystemLayerProperty = @"layer";
 - (void) updateCacheEntry:(NSUInteger)i forProperty:(NSString *)property
 {
 	NSAssert(i < OO_SYSTEM_CACHE_LENGTH,@"Invalid cache entry number");
-	NSString *key = [NSString stringWithFormat:@"%u %u",i/OO_SYSTEMS_PER_GALAXY,i%OO_SYSTEMS_PER_GALAXY];
+	NSString *key = [NSString stringWithFormat:@"%lu %lu",i/OO_SYSTEMS_PER_GALAXY,i%OO_SYSTEMS_PER_GALAXY];
 	id current = [self getProperty:property forSystemKey:key];
 	if (current == nil)
 	{
@@ -573,7 +573,7 @@ static NSString *kOOSystemLayerProperty = @"layer";
 		NSUInteger index = ([PLAYER galaxyNumber] * OO_SYSTEMS_PER_GALAXY) + s;
 		if (index >= OO_SYSTEM_CACHE_LENGTH)
 		{
-			OOLog(@"system.description.error",@"'%u' is an invalid system index for the current system. This is an internal error. Please report it.",index);
+			OOLog(@"system.description.error",@"'%lu' is an invalid system index for the current system. This is an internal error. Please report it.",(unsigned long)index);
 			return kNilRandomSeed;
 		}
 		return RandomSeedFromString([propertyCache[index] oo_stringForKey:@"random_seed"]);
